ya esta bien implementado
pero sigue subiendo muy rapido la barra

+++ MonkeyDelivery/Src/Logic/Enemy.h
@@ -8,7 +8,7 @@ protected:
 
 	double speed;
 	bool alive;
-	int scariness_;
+	double scariness_;
 	float dieTime;
 	SpawnZone zone;
 	vector<Point2D<double>> checkpoints;
